+++ soh/soh/Enhancements/Restorations/GetItemManipulation.cpp
